Loan command functional
Loan command can now be used to add a loan to the loan list of a person.
However, saving the list and the UI displaying loans has not been
achieved yet.

Refactored Loan class to accept float inputs for both interest rate
and amount, up to 2 d.p.

Refactored toString() method of Loan class to be consistent with Person
by adding a format method for Loan and changing the default toString()
method for Person to include the loan list. Made the output more human
readable.

Let's,
- Test to see if there are any unintentional bugs introduced
- Write test cases